Move FLAC split tests out of cue_test.go into flac_test.go.
